requirements.txt
fastapi==0.112.1
pydantic==2.9.1
uvicorn==0.30.6
python-dotenv==1.0.1
mysql-connector-python==9.0.0
openai==1.44.1
typing-extensions==4.12.2
starlette==0.38.5
itsdangerous==2.2.0
cryptography==43.0.1
urllib3==2.2.2
requests==2.32.3
